requirements-dev.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
django-cleanup==6.0.0
django-environ==0.9.0
Pillow==9.4.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
